requirements.txt
django==4.2.21
djangorestframework==3.16.0
python-dotenv==1.1.0
psycopg2-binary==2.9.10
openai==0.28.1
requests==2.32.3
anthropic
